rewrites api helper as pinia plugin.
