--- src/plugins/api.ts
@@ -1,13 +1,0 @@
-import { Plugin } from 'vue';
-import api from '@/helpers/api';
-
-/**
- * Adds a reference for the API helper to the Vue instance.
- */
-const plugin: Plugin = {
-  install(app) {
-    app.config.globalProperties.$api = api;
-  },
-};
-
-export default plugin;
